feat: select retro's type page
