+++ frontend/src/components/icons/Sprite.tsx
@@ -280,6 +280,25 @@ const Sprite = () => (
         fillRule="evenodd"
       />
     </symbol>
+    <symbol
+      fill="none"
+      id="blob-split-retro"
+      viewBox="0 0 31 32"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        clipRule="evenodd"
+        d="M6.36621 30.2746C14.1157 34.7636 25.2683 29.9783 28.9954 21.8554C32.7226 13.7325 30.8344 7.60134 23.0849 3.11237C15.3354 -1.37661 6.20395 -1.62868 2.47678 6.4942C-1.25039 14.6171 -1.3833 25.7857 6.36621 30.2746Z"
+        fill="#CF97FF"
+        fillRule="evenodd"
+      />
+      <path
+        clipRule="evenodd"
+        d="M21.7342 19.893C21.9809 19.4628 22.2276 19.0326 22.5359 18.4859C22.3068 18.5397 22.2011 18.5666 22.0866 18.6024C20.5095 19.0595 19.0029 18.9071 17.5403 18.1453C17.0469 17.8944 16.5095 17.6614 15.972 17.5359C12.8443 16.7831 9.76059 19.6241 10.4478 23.1104C10.8531 25.1449 12.395 26.5788 14.3333 26.8925C16.069 27.1703 18.1483 26.3279 19.1438 24.4458C20.0073 22.9222 20.8707 21.4076 21.7342 19.893L21.7342 19.893ZM8.43035 19.8752L8.43032 19.8752L8.43014 19.8748C8.13644 19.3729 7.84568 18.873 7.55492 18.3731L7.55469 18.3727C6.97328 17.3731 6.39187 16.3734 5.78697 15.3578C4.65922 13.5564 4.94996 11.2979 6.04247 9.89087C7.24952 8.32247 9.24071 7.6503 11.1879 8.29558C14.5094 9.38898 15.4434 13.5116 13.276 15.9225C12.8971 16.3437 12.4478 16.7022 11.9808 17.0159C10.6152 17.9479 9.74291 19.2027 9.37287 20.8338C9.33763 20.9324 9.30239 21.0399 9.24071 21.2729C8.92362 20.7354 8.67698 20.3053 8.43035 19.8752ZM13.3817 8.027L14.9764 8.027C15.8377 8.027 16.6924 8.02254 17.5492 8.01807C18.4104 8.01358 19.2738 8.00908 20.1482 8.00908C22.2452 7.93738 24.0073 9.35342 24.6416 11.0204C25.3465 12.8756 24.906 14.9638 23.3729 16.335C20.7562 18.6921 16.7914 17.4105 15.8575 14.2827C15.6989 13.7449 15.6196 13.1624 15.5932 12.5978C15.5051 10.9308 14.8795 9.5237 13.6989 8.36757C13.6666 8.33469 13.6357 8.30032 13.5992 8.2596C13.546 8.20043 13.4809 8.12787 13.3817 8.027Z"
+        fill="#060D16"
+        fillRule="evenodd"
+      />
+    </symbol>
     <symbol fill="none" id="blob-idea" viewBox="0 0 47 48" xmlns="http://www.w3.org/2000/svg">
       <path
         clipRule="evenodd"
@@ -306,11 +325,81 @@ const Sprite = () => (
         fillRule="evenodd"
       />
     </symbol>
+
+    <symbol
+      fill="none"
+      id="blob-arrow-right"
+      viewBox="0 0 31 32"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        clipRule="evenodd"
+        d="M6.36621 30.2746C14.1157 34.7636 25.2683 29.9783 28.9954 21.8554C32.7226 13.7325 30.8344 7.60134 23.0849 3.11237C15.3354 -1.37661 6.20395 -1.62868 2.47678 6.4942C-1.25039 14.6171 -1.3833 25.7857 6.36621 30.2746Z"
+        fill="#EC6695"
+        fillRule="evenodd"
+      />
+      <path
+        clipRule="evenodd"
+        d="M 16.7071 7.2929 C 16.3166 6.9024 15.6834 6.9024 15.2929 7.2929 C 14.9024 7.6834 14.9024 8.3166 15.2929 8.7071 L 20.5858 14 H 9 C 8.4477 14 8 14.4477 8 15 C 8 15.5523 8.4477 16 9 16 H 20.5858 L 15.2929 21.2929 C 14.9024 21.6834 14.9024 22.3166 15.2929 22.7071 C 15.6834 23.0976 16.3166 23.0976 16.7071 22.7071 L 23.7064 15.7078 C 23.7088 15.7054 23.7112 15.703 23.7136 15.7005 C 23.8901 15.5208 23.9992 15.2746 24 15.003 C 24 15.002 24 15.001 24 15 C 24 14.999 24 14.998 24 14.997 C 23.9996 14.8625 23.9727 14.7343 23.9241 14.6172 C 23.8764 14.502 23.8063 14.3938 23.7136 14.2995 C 23.7112 14.297 23.7089 14.2947 23.7065 14.2923 M 16.7071 7.2929 L 23.7065 14.2923 L 16.7071 7.2929 Z"
+        fill="currentColor"
+        fillRule="evenodd"
+      />
+    </symbol>
+
+    <symbol fill="none" id="blob-settings" viewBox="0 0 31 32" xmlns="http://www.w3.org/2000/svg">
+      <path
+        clipRule="evenodd"
+        d="M6.36621 30.2746C14.1157 34.7636 25.2683 29.9783 28.9954 21.8554C32.7226 13.7325 30.8344 7.60134 23.0849 3.11237C15.3354 -1.37661 6.20395 -1.62868 2.47678 6.4942C-1.25039 14.6171 -1.3833 25.7857 6.36621 30.2746Z"
+        fill="#F3FD58"
+        fillRule="evenodd"
+      />
+
+      <path
+        clipRule="evenodd"
+        d="M 24.913 12.8222 L 23.6581 12.6444 C 23.5685 12.2 23.3892 11.8444 23.2099 11.4889 L 23.927 10.4222 C 24.2856 9.8889 24.2856 9.0889 23.7478 8.6444 L 22.4929 7.4 C 22.0447 6.9556 21.238 6.8667 20.7002 7.2222 L 19.6245 7.9333 C 19.266 7.7556 18.8178 7.5778 18.3696 7.4 L 18.1904 6.1556 C 18.1007 5.4444 17.4733 5 16.8459 5 H 15.1428 C 14.4257 5 13.8879 5.4444 13.7983 6.1556 L 13.5294 7.4889 C 13.1708 7.6667 12.7226 7.8444 12.3641 8.0222 L 11.2885 7.3111 C 10.7507 6.9556 9.944 6.9556 9.4958 7.4889 L 8.3305 8.6444 C 7.8823 9.0889 7.7927 9.8889 8.1512 10.4222 L 8.8683 11.4889 C 8.689 11.8444 8.5098 12.2 8.4201 12.6444 L 7.1653 12.8222 C 6.4482 12.9111 6 13.5333 6 14.1556 V 15.9333 C 6 16.6444 6.4482 17.1778 7.1653 17.2667 L 8.4201 17.4444 C 8.5098 17.8 8.689 18.2444 8.8683 18.6 L 8.1512 19.6667 C 7.7927 20.2 7.7927 21 8.3305 21.4444 L 9.5854 22.6889 C 10.0336 23.1333 10.8403 23.2222 11.3781 22.8667 L 12.4537 22.1556 C 12.8123 22.3333 13.1708 22.5111 13.619 22.6 L 13.7983 23.8444 C 13.8879 24.5556 14.5153 25 15.1428 25 H 16.9355 C 17.6526 25 18.1904 24.5556 18.28 23.8444 L 18.4593 22.6 C 18.9075 22.5111 19.266 22.3333 19.6245 22.1556 L 20.7002 22.8667 C 21.238 23.2222 22.0447 23.2222 22.4929 22.6889 L 23.7478 21.4444 C 24.1959 21 24.2856 20.2 23.927 19.6667 L 23.1203 18.6 C 23.2996 18.2444 23.4789 17.8 23.5685 17.4444 L 24.8234 17.2667 C 25.5405 17.1778 25.9886 16.5556 25.9886 15.9333 V 14.1556 C 26.0783 13.5333 25.6301 12.9111 24.913 12.8222 Z M 22.3136 16.6444 C 22.1343 17.2667 21.9551 17.8 21.5965 18.3333 C 21.4172 18.6 21.5069 18.8667 21.5965 19.1333 L 22.5825 20.4667 L 21.4172 21.6222 L 20.0727 20.6444 C 19.8038 20.4667 19.5349 20.4667 19.266 20.6444 C 18.7282 20.9111 18.1904 21.1778 17.5629 21.3556 C 17.294 21.4444 17.0251 21.6222 17.0251 21.9778 L 16.7562 23.6667 H 15.1428 L 14.8739 21.9778 C 14.7842 21.7111 14.605 21.4444 14.3361 21.3556 C 13.7983 21.1778 13.1708 21 12.7226 20.6444 C 12.4537 20.4667 12.1848 20.5556 11.9159 20.6444 L 10.5714 21.6222 L 9.4061 20.4667 L 10.3921 19.1333 C 10.5714 18.8667 10.5714 18.6 10.3921 18.3333 C 10.0336 17.8 9.8543 17.2667 9.675 16.6444 C 9.5854 16.3778 9.4061 16.2 9.0476 16.1111 L 7.3445 15.8444 V 14.2444 L 9.0476 13.9778 C 9.3165 13.8889 9.5854 13.7111 9.675 13.4444 C 9.8543 12.8222 10.0336 12.2889 10.3921 11.7556 C 10.5714 11.4889 10.4818 11.2222 10.3921 10.9556 L 9.4061 9.6222 L 10.5714 8.4667 L 11.9159 9.4444 C 12.1848 9.6222 12.4537 9.6222 12.7226 9.4444 C 13.2605 9.1778 13.7983 8.9111 14.3361 8.8222 C 14.605 8.7333 14.7842 8.5556 14.8739 8.2 L 15.1428 6.5111 H 16.7562 L 17.0251 8.2 C 17.1148 8.4667 17.294 8.7333 17.5629 8.8222 C 18.1904 9 18.7282 9.1778 19.266 9.5333 C 19.5349 9.7111 19.8038 9.6222 20.0727 9.5333 L 21.4172 8.5556 L 22.5825 9.7111 L 21.5965 11.0444 C 21.4172 11.3111 21.4172 11.5778 21.5965 11.8444 C 21.8654 12.3778 22.1343 12.9111 22.3136 13.5333 C 22.4032 13.8 22.5825 14.0667 22.941 14.0667 L 24.6441 14.3333 V 15.9333 L 22.941 16.2 C 22.6721 16.2 22.4032 16.3778 22.3136 16.6444 Z"
+        fill="#434D5A"
+        fillRule="evenodd"
+      />
+      <path
+        clipRule="evenodd"
+        d="M 16.0392 10.7773 C 13.619 10.7773 11.7367 12.7329 11.7367 15.044 C 11.7367 17.444 13.7087 19.3107 16.0392 19.3107 C 18.3697 19.3107 20.3417 17.3551 20.3417 15.044 C 20.3417 12.644 18.4593 10.7773 16.0392 10.7773 Z M 16.0392 18.0662 C 14.3361 18.0662 12.9916 16.7329 12.9916 15.044 C 12.9916 13.3551 14.3361 12.0218 16.0392 12.0218 C 17.7422 12.0218 19.0868 13.3551 19.0868 15.044 C 19.0868 16.7329 17.7422 18.0662 16.0392 18.0662 Z"
+        fill="#434D5A"
+        fillRule="evenodd"
+      />
+
+      <defs>
+        <clipPath id="clip0_3271_122787">
+          <rect fill="white" height="18" transform="translate(20.05176 7.55322)" width="18" />
+        </clipPath>
+      </defs>
+    </symbol>
+
     <symbol fill="none" id="blob-team" viewBox="0 0 31 32" xmlns="http://www.w3.org/2000/svg">
       <path
         clipRule="evenodd"
         d="M6.36621 30.2746C14.1157 34.7636 25.2683 29.9783 28.9954 21.8554C32.7226 13.7325 30.8344 7.60134 23.0849 3.11237C15.3354 -1.37661 6.20395 -1.62868 2.47678 6.4942C-1.25039 14.6171 -1.3833 25.7857 6.36621 30.2746Z"
         fill="#73FBF7"
+        fillRule="evenodd"
+      />
+
+      <path
+        clipRule="evenodd"
+        d="M13.1233 13.0175C13.1233 11.9524 13.9867 11.0889 15.0518 11.0889C16.1169 11.0889 16.9804 11.9524 16.9804 13.0175C16.9804 14.0826 16.1169 14.9461 15.0518 14.9461C13.9867 14.9461 13.1233 14.0826 13.1233 13.0175ZM15.0518 9.80322C13.2766 9.80322 11.8375 11.2423 11.8375 13.0175C11.8375 14.7927 13.2766 16.2318 15.0518 16.2318C16.827 16.2318 18.2661 14.7927 18.2661 13.0175C18.2661 11.2423 16.827 9.80322 15.0518 9.80322ZM11.0353 9.90696C11.3792 9.8189 11.7294 10.0263 11.8175 10.3703C11.9055 10.7142 11.6981 11.0644 11.3542 11.1525C10.9393 11.2587 10.5716 11.5 10.3091 11.8383C10.0465 12.1765 9.90396 12.5926 9.90396 13.0208C9.90396 13.449 10.0465 13.8651 10.3091 14.2033C10.5716 14.5416 10.9393 14.7829 11.3542 14.8891C11.6981 14.9772 11.9055 15.3274 11.8175 15.6713C11.7294 16.0153 11.3792 16.2227 11.0353 16.1346C10.3439 15.9576 9.73104 15.5555 9.29341 14.9917C8.85579 14.4279 8.61825 13.7345 8.61825 13.0208C8.61825 12.3071 8.85579 11.6137 9.29341 11.0499C9.73104 10.4861 10.3439 10.084 11.0353 9.90696ZM8.46247 17.6219C8.80624 17.5331 9.15687 17.7399 9.24563 18.0836C9.33439 18.4274 9.12766 18.778 8.7839 18.8668C8.37016 18.9736 8.00361 19.2148 7.74181 19.5525C7.48005 19.8902 7.33783 20.3054 7.33747 20.7327L7.33747 22.0179C7.33747 22.3729 7.04965 22.6608 6.69461 22.6608C6.33957 22.6608 6.05176 22.3729 6.05176 22.0179L6.05176 20.7317C6.05229 20.0195 6.28933 19.3277 6.72566 18.7648C7.16199 18.202 7.7729 17.7999 8.46247 17.6219ZM10.2075 18.4585C10.8102 17.8557 11.6278 17.5171 12.4803 17.5171H17.6231C18.4756 17.5171 19.2932 17.8557 19.896 18.4585C20.4988 19.0613 20.8374 19.8789 20.8374 20.7314V22.0171C20.8374 22.3721 20.5496 22.6599 20.1946 22.6599C19.8395 22.6599 19.5517 22.3721 19.5517 22.0171V20.7314C19.5517 20.2199 19.3485 19.7293 18.9869 19.3677C18.6252 19.006 18.1346 18.8028 17.6231 18.8028H12.4803C11.9688 18.8028 11.4783 19.006 11.1166 19.3677C10.7549 19.7293 10.5517 20.2199 10.5517 20.7314V22.0171C10.5517 22.3721 10.2639 22.6599 9.90887 22.6599C9.55383 22.6599 9.26601 22.3721 9.26601 22.0171V20.7314C9.26601 19.8789 9.60466 19.0613 10.2075 18.4585ZM20.8579 18.0836C20.9467 17.7399 21.2973 17.5331 21.6411 17.6219C22.3307 17.7999 22.9416 18.202 23.3779 18.7648C23.8142 19.3277 24.0513 20.0195 24.0518 20.7317V22.0179C24.0518 22.3729 23.764 22.6608 23.4089 22.6608C23.0539 22.6608 22.7661 22.3729 22.7661 22.0179V20.7327C22.7657 20.3054 22.6235 19.8902 22.3617 19.5525C22.0999 19.2148 21.7334 18.9736 21.3197 18.8668C20.9759 18.778 20.7692 18.4274 20.8579 18.0836ZM19.0683 9.90696C18.7243 9.8189 18.3741 10.0263 18.2861 10.3703C18.198 10.7142 18.4054 11.0644 18.7494 11.1525C19.1642 11.2587 19.5319 11.5 19.7945 11.8383C20.0571 12.1765 20.1996 12.5926 20.1996 13.0208C20.1996 13.449 20.0571 13.8651 19.7945 14.2033C19.5319 14.5416 19.1642 14.7829 18.7494 14.8891C18.4054 14.9772 18.198 15.3274 18.2861 15.6713C18.3741 16.0153 18.7243 16.2227 19.0683 16.1346C19.7597 15.9576 20.3725 15.5555 20.8101 14.9917C21.2478 14.4279 21.4853 13.7345 21.4853 13.0208C21.4853 12.3071 21.2478 11.6137 20.8101 11.0499C20.3725 10.4861 19.7597 10.084 19.0683 9.90696Z"
+        fill="#060D16"
+        fillRule="evenodd"
+      />
+
+      <defs>
+        <clipPath id="clip0_3271_122787">
+          <rect fill="white" height="18" transform="translate(6.05176 7.55322)" width="18" />
+        </clipPath>
+      </defs>
+    </symbol>
+    <symbol fill="none" id="blob-team-retro" viewBox="0 0 31 32" xmlns="http://www.w3.org/2000/svg">
+      <path
+        clipRule="evenodd"
+        d="M6.36621 30.2746C14.1157 34.7636 25.2683 29.9783 28.9954 21.8554C32.7226 13.7325 30.8344 7.60134 23.0849 3.11237C15.3354 -1.37661 6.20395 -1.62868 2.47678 6.4942C-1.25039 14.6171 -1.3833 25.7857 6.36621 30.2746Z"
+        fill="#AEF6CD"
         fillRule="evenodd"
       />
