+++ frontend/src/styles/stitches/partials/sizes.ts
@@ -42,6 +42,7 @@ const sizes = {
   237: '14.813rem',
   260: '16.25rem',
   300: '18.75rem',
+  308: '19.25rem',
   353: '22.063rem',
   360: '22.5rem',
   384: '24rem',
